Add style to page
Created CSS stylesheet. There are some lingering issues that make the
page scale incorrectly to the window. The images in the image bar do
not shrink to accommodate the window like the hero bar does. Margins are
static in the quote bar, leading to unattractive formatting when shrinking
the window.